backend/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
scikit-learn
numpy
sentence-transformers
flake8
pytest-django
